MXRoom: Fix detection of the end of the back pagination.
End of pagination is now detected when returned chunk is empty and both tokens (start/end) are equal.

CAUTION: At MXRoom level, a back pagination may be complete whereas no new event has been added. Indeed the server may returned an empty chunk with updated start and end tokens when pagination takes place in a moment where the user was not in the room.
The developer should handle this case at application level in order to loop on a new pagination request or not.