Unread Count Handling
At the end of the initial sync, the ios client may compute the unread count for each room thanks to the read receipt. This count is wrong if the last read event is not present in local storage.
To solve this pb, the server should be able to send the unread count for each room else the client will have to back paginate until the last read event id...

Previously the ios client marked locally the latest event of each room as read in order to reset locally unread counter.
We change here this behavior to identify the rooms with unread messages even if the unread count is incorrect.