Room summary: bug fix: In case of state event redaction, the last message may be empty after a clear cache.
The reason was that the code arose a room initial sync when it sees the redaction event from the full initial sync response after the clear cache.
This creates some race conditions.

This room initial sync is useless in this case because we get already sanitised data from the full initial sync. No need to claim for fresh data from the HS.