Crypto: Do not block the main thread when retrying the decryption of an event.
Retrying happens when we receive the key afterwards because of a late room key or because another user's device answers to our key share request.
In that case, as we are already in an asynchronously case, we do not need to come back to and block the main thread to decrypt the event as opposed to https://github.com/matrix-org/matrix-ios-sdk/issues/205.