MXFileStore: BF: The MXFileStore metadata (containing eventStremToken) can be not stored in files.
When there is no eventStremToken, the MXFileStore is reset on the next start, requiring a full initial sync :/

This happened when there are more 2 pending commits. the dispatch saving the copy of the MXFileStoreMetaData data was skipped and the copied data was lost.
As other component in MXFileStore, do not do copy but reuse the same instance.

Found with the checkRoomAccountDataTags test.