Bug Fix - Direct Chat: a room was marked as direct by mistake when I joined it.
We decided here to disable the heuristic method applied on invite when the 'is_direct' tag was missing in the invite event.

Indeed if the 'is_direct' tag was missing, we were used to apply heuristics to decide whether to consider it a DM (given it may have come from a client that doesn't know about m.direct).
This introduced some wrong direct chats (for example when the user joined a room whereas the room creation and the invitations were done separately).