MXHTTPClient: Enable the certificate pinning mode by default as soon as some certificates are present in the application bundle.
In the previous code, we let AFSecurityPolicy try to retrieve certificate in the AFNetworking Bundle. But, this bundle is not equal to the app Bundle. So certificates were not retrieved.
BTW we do not allow InvalidCertificates anymore when the certificates pinning mode is enabled.

MXRestClient: support multiple SSL pinning modes (none/public key/certificate) too.