Bug Fix : Some room members count are wrong after clearing the cache
The bug was located in [MXRoomState copyWithZone:]. Indeed the item `membersCount` was not copied correctly whereas this is an instance of MXRoomMembersCount class.

The "back state" created thanks to the copy here: https://github.com/matrix-org/matrix-ios-sdk/blob/6027fd5b2b644f2ebeb1049cdb80a318d7db2273/MatrixSDK/Data/MXRoomState.m#L126 was using the same membersCount instance as the provided "live state". The content of this membersCount was updated in case of membership events in the back pagination. The live state was then corrupted.
According to my investigation, the corrupted state was not pushed to the store, except perhaps if new membership events were received in the live stream.