MXSession: Fix a race conditions that prevented MXSession from actually being paused
The bug exists since lazy loading implementation and the a-synchronousness it added.
When the race appears, the session does not pause which leads to bg kills on iOS (https://github.com/vector-im/element-ios/issues/3761).

This commit fixes also the resume operation that never returned if the app was in running state (which was the state with the race condition).