MXBackgroundSyncService: Disable usage of MXSession store data
It uses too much RAM when running within an extension.
The culprit is the deserialisation of all messages of a room.

Note that we have the same out of memory issue when loading the room state of a big room but this is not new. Related issue: https://github.com/vector-im/element-ios/issues/3868