BugFix - MXRoomSummary: directUserId may be missing (null) for a direct chat if it was joined on another device.
I observed a direct chat for which the field directUserId was null in its room summary. NOK

This field of the room summaries is updated currently in 3 cases (at the sdk level during the sync):
- initial sync
- new invites in the server sync response
- an update of the direct chats (in the account data)

I remembered join this direct chat on a web client, I realized the room summary was not correctly updated on iOS because the room was not present in the rooms list when we processed the new direct chat list. Indeed the account data (in which are the directs) are handled before the joined room list...
In conclusion: directUserId is missing (null) for a direct chat after the sync if it was joined on another device whereas the ios session was offline.