MXRoomState/MXRoomMembers: Fix memory leak and copying
`MXRoomState` and `MXRoomMembers` held mutual strong references to each other via
the `members` property and the `state` ivar, respectively. This led to a memory leak
preventing both objects from being deallocated.

Additionally, the leak was masqueraded by a bug in `MXRoomState`'s `copyWithZone:`
method. The latter copied the `MXRoomMembers` object but didn't update its `state`
ivar to point to the `MXRoomState` copy. As a result, copying an `MXRoomState` object
did *not* introduce another leak.

The present commit addresses both issues. First, the `state` ivar on `MXRoomMembers`
is made weak. This seems safe because `MXRoomState` retains its `members` property
and `MXRoomMembers` should not exist on its own without a room state owning it.

Secondly, a new method `copyWithZone:andState:` is introduced on `MXRoomMembers`. The
latter copies the receiver and updates the `state` ivar to the specified `MXRoomState`.

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>