MXDeviceListOperation: Fix memory leak
This breaks a strong reference cycle between `MXDeviceListOperation` and
`MXDeviceListOperationsPool`. When an operation is added to a pool, both objects
hold mutual strong references to each other by means of the `pool` and `_operations`
ivars. This prevents them from being deallocated when the operation has finished.

The cycle is broken by weakifying the `pool` ivar in `MXDeviceListOperation`.
This is safe because the operation should always be managed by an operation pool
and the latter is retained in `MXDeviceList`.

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>