MXDeviceList: Fix memory leak
When creating an `MXDeviceListOperation` in `MXDeviceList`, the block passed into
the `success` parameter captures a strong reference to the operation. Since the
`success` block is stored with a strong reference in `MXDeviceListOperation`, this
creates a strong reference cycle that prevents the operation from being deallocated.

This commit breaks the cycle by passing a weak reference to the operation into the
`success` block. Using a weak reference is safe because, as mentioned above, the
block is held as a strong reference by the operation itself.

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>