MXRoomSummary: Do not decrypt synchronously lastMessageEvent anymore
lastMessageEvent is required by room list. It took 1.3 seconds on the main thread to get it on my account (600 rooms). Now there are asynchronously preloaded and decrypted when the SDK load.

One test has been fixed. As events are now decrypted only upstream and no more on demand, we need to trigger a decrypt to mimic this upstream decryption.