MXCrypto: eventDeviceInfo: Do not synchronise the operation anymore with the decryption queue.
A single decryption takes 5ms on an iPhoneX. It was useless to synchronise this method with the busy decryption thread. There is no relationship.

Note that this operation uses 0.5ms. It must not be called from the main thread.
MatrixKit uses this method to build the timeline. It uses on a dedicated queue.