MXSession: Fix app that can fail to resume
vector-im/element-ios/issues/4417

The root cause of the issue is a race condition in the MXSession pause management between the sdk state machine and the kit state machine.

The SDK was blocked and dead in MXSessionStateSyncInProgress state. The kit thought it does not need to resume it on app resume.

This fix is very related on how the kit and the sdk work together. But it could make sense in a more general manner.

I was able to reproduce the issue ("[MXSession] The connection has been cancelled.") by adding a delay of 25s in /sync response. Then, if you background the app just after starting it, you will hit this case.